Use `intptr_t` for counting SolidSegs
This should be more correct. Thanks @rfomin!

+++ src/doom/r_bsp.c
@@ -566,7 +566,7 @@ void R_Subsector (int num)
     }
 
     // [JN] Count solidsegs limit.
-    CRLData.numsolidsegs = (int)(newend - solidsegs);
+    CRLData.numsolidsegs = (intptr_t)(newend - solidsegs);
 
     // check for solidsegs overflow - extremely unsatisfactory!
     // [JN] CRL - Do not quit with I_Error, print in-game warning instead.
